refactor: restructure src/vcs/ into pr/, commit/, and auth/ subdirectories
Organize the flat 15-file vcs/ directory into logical subdirectories
while preserving all exports through barrel index files, so existing
imports via src/vcs/index.js continue to work unchanged.
