+++ src/vcs/auth/index.ts
@@ -1,0 +1,1 @@
+export { GitHubAppTokenManager } from "./github-app-token-manager.js";
