+++ src/vcs/commit/index.ts
@@ -1,0 +1,13 @@
+export {
+  createCommitStrategy,
+  createTokenManager,
+} from "./commit-strategy-selector.js";
+export {
+  GraphQLCommitStrategy,
+  validateSafeBranchName,
+  MAX_PAYLOAD_SIZE,
+  SAFE_BRANCH_NAME_PATTERN,
+} from "./graphql-commit-strategy.js";
+export { GitCommitStrategy } from "./git-commit-strategy.js";
+export { FileModeFixupCommitStrategy } from "./file-mode-fixup-commit-strategy.js";
+export type { GhApiClientFactory } from "./file-mode-fixup-commit-strategy.js";
