+++ src/vcs/index.ts
@@ -16,17 +16,14 @@ export type {
 export type { GitOpsOptions } from "./git-ops.js";
 
 // Commit strategies
-export {
-  createCommitStrategy,
-  createTokenManager,
-} from "./commit-strategy-selector.js";
-export { FileModeFixupCommitStrategy } from "./file-mode-fixup-commit-strategy.js";
+export { createCommitStrategy, createTokenManager } from "./commit/index.js";
+export { FileModeFixupCommitStrategy } from "./commit/index.js";
 // Token manager type (concrete class used as type by sync module)
-export type { GitHubAppTokenManager } from "./github-app-token-manager.js";
+export type { GitHubAppTokenManager } from "./auth/index.js";
 // Git operations
 export { GitOps } from "./git-ops.js";
 export { AuthenticatedGitOps } from "./authenticated-git-ops.js";
 // PR strategy factory
-export { createPRStrategy } from "./pr-strategy-factory.js";
+export { createPRStrategy } from "./pr/index.js";
 // PR creation and merge
-export { createPR, mergePR } from "./pr-creator.js";
+export { createPR, mergePR } from "./pr/index.js";
