+++ src/vcs/pr/pr-strategy-factory.ts
@@ -3,13 +3,13 @@ import {
   isGitHubRepo,
   isAzureDevOpsRepo,
   isGitLabRepo,
-} from "../repo/index.js";
-import type { IPRStrategy } from "./types.js";
-import { SyncError } from "../shared/errors.js";
+} from "../../repo/index.js";
+import type { IPRStrategy } from "../types.js";
+import { SyncError } from "../../shared/errors.js";
 import { GitHubPRStrategy } from "./github-pr-strategy.js";
 import { AdoPRStrategy } from "./ado-pr-strategy.js";
 import { GitLabPRStrategy } from "./gitlab-pr-strategy.js";
-import type { ICommandExecutor } from "../shared/command-executor.js";
+import type { ICommandExecutor } from "../../shared/command-executor.js";
 import type { IPRStrategyLogger } from "./pr-strategy.js";
 
 export function createPRStrategy(
