+++ test/unit/sync/auth-options-builder.test.ts
@@ -2,7 +2,7 @@ import { test, describe } from "node:test";
 import { strict as assert } from "node:assert";
 import { AuthOptionsBuilder } from "../../../src/sync/auth-options-builder.js";
 import type { GitHubRepoInfo } from "../../../src/repo/index.js";
-import type { GitHubAppTokenManager } from "../../../src/vcs/github-app-token-manager.js";
+import type { GitHubAppTokenManager } from "../../../src/vcs/auth/github-app-token-manager.js";
 
 /** Mock token manager - only needs getTokenForRepo method */
 type MockTokenManager = Pick<GitHubAppTokenManager, "getTokenForRepo">;
