+++ test/unit/sync/commit-message.test.ts
@@ -1,7 +1,7 @@
 import { test, describe } from "node:test";
 import { strict as assert } from "node:assert";
 import { formatCommitMessage } from "../../../src/sync/commit-message.js";
-import type { FileAction } from "../../../src/vcs/pr-creator.js";
+import type { FileAction } from "../../../src/vcs/pr/pr-creator.js";
 
 describe("formatCommitMessage", () => {
   test("returns single file message for one changed file", () => {
