+++ test/unit/vcs/git-commit-strategy.test.ts
@@ -3,7 +3,7 @@ import { strict as assert } from "node:assert";
 import { mkdirSync, rmSync, existsSync } from "node:fs";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
-import { GitCommitStrategy } from "../../../src/vcs/git-commit-strategy.js";
+import { GitCommitStrategy } from "../../../src/vcs/commit/git-commit-strategy.js";
 import { GitHubRepoInfo } from "../../../src/repo/index.js";
 import { CommitOptions } from "../../../src/vcs/types.js";
 import type { INetworkGitOps } from "../../../src/vcs/types.js";
