+++ test/unit/vcs/github-pr-strategy.test.ts
@@ -3,8 +3,8 @@ import { strict as assert } from "node:assert";
 import { mkdirSync, rmSync, existsSync } from "node:fs";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
-import { GitHubPRStrategy } from "../../../src/vcs/github-pr-strategy.js";
-import { PRWorkflowExecutor } from "../../../src/vcs/pr-strategy.js";
+import { GitHubPRStrategy } from "../../../src/vcs/pr/github-pr-strategy.js";
+import { PRWorkflowExecutor } from "../../../src/vcs/pr/pr-strategy.js";
 import { GitHubRepoInfo } from "../../../src/repo/index.js";
 import type { PRStrategyOptions } from "../../../src/vcs/types.js";
 import {
