+++ test/unit/vcs/pr-strategy-factory.test.ts
@@ -1,6 +1,6 @@
 import { test, describe } from "node:test";
 import { strict as assert } from "node:assert";
-import { createPRStrategy } from "../../../src/vcs/pr-strategy-factory.js";
+import { createPRStrategy } from "../../../src/vcs/pr/pr-strategy-factory.js";
 import type { GitHubRepoInfo } from "../../../src/repo/index.js";
 import type { AzureDevOpsRepoInfo } from "../../../src/repo/index.js";
 import type { GitLabRepoInfo } from "../../../src/repo/index.js";
